Ignore unknown fields when unmarshaling config files
Because of the asynchronous nature of application upgrades in
Kubernetes, there exists a time where the contents of the ConfigMap will
differ from what existing Antrea pods are able to marshal successfully.
This leads to those pods crashing until they're replaced by their parent
resource.

This fix ensures that field additions do not break existing pods during
that upgrade process.

Note that field removal or changes to the valid values for a field will
still exhibit the same crashing behavior until the pod can be replaced.